+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - use "return" command for conditionals instead of data storage when using supported pack format
+- update latest datapack format to 61
 
 ### Removed